requirements.txt
python-dotenv>=1.0.1
Pillow>=10.4.0
numpy>=2.1.0
scikit-learn>=1.5.1
joblib>=1.4.2
imagehash>=4.3.1
pyzipper>=0.3.6
pytest>=8.3.2
